docs: relocate data-grid visibility demos to examples
