+++ docs/content/blog/v1-5-release.mdx
@@ -10,8 +10,6 @@ timeToRead: 3
 tag: Release
 ---
 
-import { DEMOS as VisibilityDemos } from '@/app/docs/data-grid/[slug]/components/visibility/static'
-
 This release provides new features to the Data Grid along with general bugfixes.
 
 ## Overview
@@ -58,7 +56,7 @@ This is achieved by setting the `features.visibility` prop to `true` or by utili
 
 > **Note**: We have updated the background color of Data Grid Popovers used for features to improve contrast and readability.
 
-<CodePreview {...VisibilityDemos.basic} />
+<Example path="data-grid/visibility" demo="basic" />
 
 #### New Context Values
