+++ docs/examples/data-grid/visibility/index.tsx
@@ -3,17 +3,12 @@ import { HiddenDemo } from './hidden.demo'
 
 export const DEMOS = {
   basic: {
-    id: 'visibility.basic',
     preview: <BasicDemo />,
-    context: 'data-grid',
   },
   hidden: {
-    id: 'visibility.hidden',
     preview: <HiddenDemo />,
-    context: 'data-grid',
   },
   def: {
-    id: 'visibility.def',
-    context: 'data-grid',
+    preview: null,
   },
 }
